Moved some stuff about in server test proj
